requirements.txt
ffmpeg-python==0.2.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.8
pymongo==4.6.1
sqlalchemy==2.0.25
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pytesseract==0.3.10
tmdbsimple==2.9.1 